pyproject.toml
[tool.poetry]
name = "bored-bot"
version = "0.1.0"
description = "A bot that does stuff awfully"
authors = ["mackdroid <68280146+mackdroid@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "~3.9"
requests = "^2.27.1"
PyNaCl = "^1.5.0"
alt-profanity-check = "^1.0.2"
asyncio = "^3.4.3"
youtube_dl = "^2021.12.17"
nextcord = "^2.0.0-alpha.6"
Pillow = "^9.0.1"
tinydb = "^4.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
